refactor(docker): refactor docker-compose | add scripts for testing setup

+++ scripts/setup.sh
@@ -1,11 +1,21 @@
+echo "Starting Setup..."
+
+if ! (docker ps | grep -q "postgres-container-wallet") || ! (docker ps | grep -q "mongo-container-wallet") || ! (docker ps | grep -q "wallet-service-container"); then
+    docker-compose -f docker/docker-compose.yml down
+    docker-compose -f docker/docker-compose.yml up -d
+    sleep 5 
+fi
 echo "Init Containers! 🚀"
 
-docker-compose -f docker/docker-compose.yml up -d
+url="http://localhost:6000/ok"
+http_status=$(curl -o /dev/null -I -sw "%{http_code}" $url)
+if [ $http_status -eq 200 ]; then
+    echo "Wallet Service is OK! ✅"
+else
+    echo "Wallet Service Failed! ❌"
+fi
 
-sleep 5 
-echo "Test API:"
-curl http://localhost:6000/health
+echo "Postgres Setup..."
+./scripts/postgres-setup.sh
 
-
-sleep 2
-docker-compose -f docker/docker-compose.yml down
+echo "Setup Completed! ✅"
